GlobalRates_web/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.1.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1